refactor(uns): move uns out of evm folder
