+++ src/core/Solana/resolveSolanaAddress.ts
@@ -1,5 +1,5 @@
 import { ChainType } from '@lifi/types'
-import { resolveUNSAddress } from '../EVM/uns/resolveUNSAddress.js'
+import { resolveUNSAddress } from '../uns/resolveUNSAddress.js'
 import { getSNSAddress } from './getSNSAddress.js'
 
 export async function resolveSolanaAddress(
